+++ vortex-file/src/v2/exec/mod.rs
@@ -1,6 +1,8 @@
+pub mod inline;
+pub mod tokio;
+
 use futures_util::future::BoxFuture;
 use futures_util::stream::BoxStream;
-use futures_util::StreamExt;
 use vortex_array::ArrayData;
 use vortex_error::VortexResult;
 
@@ -10,16 +12,4 @@ pub trait ExecDriver {
         &self,
         stream: BoxStream<'static, BoxFuture<'static, VortexResult<ArrayData>>>,
     ) -> BoxStream<'static, VortexResult<ArrayData>>;
-}
-
-/// An execution driver that awaits the futures inline using the caller's runtime.
-pub struct InlineDriver;
-
-impl ExecDriver for InlineDriver {
-    fn drive(
-        &self,
-        stream: BoxStream<'static, BoxFuture<'static, VortexResult<ArrayData>>>,
-    ) -> BoxStream<'static, VortexResult<ArrayData>> {
-        stream.then(|future| future).boxed()
-    }
 }
