+++ vortex-file/src/v2/exec/tokio.rs
@@ -1,0 +1,27 @@
+use futures_util::future::BoxFuture;
+use futures_util::stream::BoxStream;
+use futures_util::StreamExt;
+use tokio::runtime::Handle;
+use vortex_array::ArrayData;
+use vortex_error::{vortex_err, VortexResult};
+
+use crate::v2::exec::ExecDriver;
+
+/// An execution driver that spawns the futures onto a Tokio runtime.
+pub struct TokioDriver(pub Handle);
+
+impl ExecDriver for TokioDriver {
+    fn drive(
+        &self,
+        stream: BoxStream<'static, BoxFuture<'static, VortexResult<ArrayData>>>,
+    ) -> BoxStream<'static, VortexResult<ArrayData>> {
+        let handle = self.0.clone();
+        stream
+            .then(move |future| handle.spawn(future))
+            .map(|result| match result {
+                Ok(result) => result,
+                Err(e) => Err(vortex_err!("Failed to join Tokio result {}", e)),
+            })
+            .boxed()
+    }
+}
